+++ activesupport/CHANGELOG.md
@@ -1,3 +1,12 @@
+*   Fix not calling `#default` on `HashWithIndifferentAcess#to_hash` when only
+    `default_proc` is set, which could raise.
+
+    *Simon Eskildsen*
+
+*   Fix setting `default_proc` on `HashWithIndifferentAccess#dup`
+
+    *Simon Eskildsen*
+
 *   Fix a range of values for parameters of the Time#change
 
     *Nikolay Kondratyev*